# frontend/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger
  version: 1.0.0
...
paths:
...
  /basket:
    get:
      tags:
        - basket
      description: 'Get items in basket'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductShort'
    post:
      tags:
        - basket
      description: 'Add item to basket'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              xml:
                name: item
              properties:
                id:
                  type: number
                  example: 123
                count:
                  type: number
                  example: 5
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basket'
    delete:
      tags:
        - basket
      description: 'Remove item from basket'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              xml:
                name: item
              properties:
                id:
                  type: number
                  example: 123
                count:
                  type: number
                  example: 5
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Basket'
...
components:
  schemas:
...
    ProductShort:
      required:
        - id
        - category
        - title
        - price
        - count
      type: object
      xml:
        name: productShort
      properties:
        id:
          type: number
          example: 123
        category:
          type: number
          example: 55
        price:
          type: number
          example: 500.67
        count:
          type: integer
          example: 12
        date:
          type: string
          example: 'Thu Feb 09 2023 21:39:52 GMT+0100 (Central European Standard Time)'
        title:
          type: string
          example: 'video card'
        description:
          type: string
          example: 'description of the product'
        freeDelivery:
          type: boolean
          example: true
        images:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Image'
        tags:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        reviews:
          type: integer
          example: 5
        rating:
          type: number
          example: 4.6
...
    OrderProduct:
      allOf:
        - $ref: '#/components/schemas/ProductShort'
        - description: 'count - для количества товара в корзине и price их общая цена на момент покупки'
        - type: object
        - xml:
            name: orderProduct
...
    Basket:
      type: array
      xml:
        name: basket
      items:
        $ref: '#/components/schemas/OrderProduct'
...
    Tag:
      type: object
      xml:
        name: tag
      properties:
        id:
          type: number
          example: 12
        name:
          type: string
          example: 'Gaming'
    Image:
      type: object
      xml:
        name: image
      properties:
        src:
          type: string
          example: '/3.png'
        alt:
          type: string
          example: 'Image alt string'